+++ Firestore/CHANGELOG.md
@@ -3,6 +3,8 @@
   has been removed. See
   https://firebase.google.com/docs/ios/swift-migration for migration
   instructions.
+- **Breaking change**: LRU Garbage Collector is now the default GC for memory cache, eager GC is now
+  opt-in instead of the default one.
 
 # 10.25.0
 - [fixed] Allow blob of data with zero length. (#11773, #12620)